Made the code into a framework. Seperated into various files
